requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
pkg-resources==0.0.0
pyee==9.0.4
typing-extensions==4.3.0
